Move script_builder to gui
